Prompts/requirements.txt
langchain
langchain-groq
groq
streamlit
python-dotenv
langchain-core  